requirements.txt
pandas
python-dotenv
openai
numpy
scipy
pyyaml